// src/audio/clip/types.ts
export interface AudioMetadata {
	title?: string;
	artist?: string;
	album?: string;
	trackNumber?: number;
	year?: number;
	genre?: string;
	duration?: number;
	sampleRate?: number;
	channels?: number;
	bitrate?: number;
	codec?: string;
	picture?: {
		data: ArrayBuffer;
		mimeType: string;
	};
}